more cyclic dependency removal
