+++ src/vs/workbench/contrib/chat/common/chatEditingService.ts
@@ -8,12 +8,15 @@ import { Event } from '../../../../base/common/event.js';
 import { IDisposable } from '../../../../base/common/lifecycle.js';
 import { IObservable, IReader } from '../../../../base/common/observable.js';
 import { URI } from '../../../../base/common/uri.js';
+import { StringEdit } from '../../../../editor/common/core/edits/stringEdit.js';
 import { TextEdit } from '../../../../editor/common/languages.js';
+import { ITextModel } from '../../../../editor/common/model.js';
 import { localize } from '../../../../nls.js';
 import { RawContextKey } from '../../../../platform/contextkey/common/contextkey.js';
 import { createDecorator } from '../../../../platform/instantiation/common/instantiation.js';
 import { IEditorPane } from '../../../common/editor.js';
 import { ICellEditOperation } from '../../notebook/common/notebookCommon.js';
+import { IChatAgentResult } from './chatAgents.js';
 import { ChatModel, IChatResponseModel } from './chatModel.js';
 
 export const IChatEditingService = createDecorator<IChatEditingService>('chatEditingService');
@@ -79,6 +82,25 @@ export interface IStreamingEdits {
 
 export const chatEditingSnapshotScheme = 'chat-editing-snapshot-text-model';
 
+export interface IModifiedEntryTelemetryInfo {
+	readonly agentId: string | undefined;
+	readonly command: string | undefined;
+	readonly sessionId: string;
+	readonly requestId: string;
+	readonly result: IChatAgentResult | undefined;
+}
+
+export interface ISnapshotEntry {
+	readonly resource: URI;
+	readonly languageId: string;
+	readonly snapshotUri: URI;
+	readonly original: string;
+	readonly current: string;
+	readonly originalToCurrentEdit: StringEdit;
+	readonly state: ModifiedFileEntryState;
+	telemetryInfo: IModifiedEntryTelemetryInfo;
+}
+
 export interface IChatEditingSession extends IDisposable {
 	readonly isGlobalEditingSession: boolean;
 	readonly chatSessionId: string;
@@ -99,6 +121,10 @@ export interface IChatEditingSession extends IDisposable {
 	 * @param uri File in the workspace
 	 */
 	getSnapshotUri(requestId: string, uri: URI, stopId: string | undefined): URI | undefined;
+
+	getSnapshotModel(requestId: string, undoStop: string | undefined, snapshotUri: URI): Promise<ITextModel | null>;
+
+	getSnapshot(requestId: string, undoStop: string | undefined, snapshotUri: URI): ISnapshotEntry | undefined;
 
 	/**
 	 * Will lead to this object getting disposed
